refactor: rename classes to be sensible Rename departures_at_stop to departure, nearby_stations to stop, searched_stations to station
