+++ lib/services/vbb_api.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'package:departures/services/departures_at_stop.dart';
-import 'package:departures/services/nearby_stations.dart';
-import 'package:departures/services/searched_stations.dart';
+import 'package:departures/services/departure.dart';
+import 'package:departures/services/stop.dart';
+import 'package:departures/services/station.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:http/http.dart' as http;
@@ -31,7 +31,7 @@ class VbbApi {
     }
   }
 
-  static Future<List<NearbyStation>> getNearbyStations(double latitude, double longitude, BuildContext context) async {
+  static Future<List<Stop>> getNearbyStations(double latitude, double longitude, BuildContext context) async {
     final String host = await _getMainApiHost();
 
     final response = await http.get(
@@ -50,7 +50,7 @@ class VbbApi {
 
     if (response.statusCode == 200) {
       Iterable stations = json.decode(response.body);
-      List<NearbyStation> nearbyStations = List<NearbyStation>.from(stations.map((model) => NearbyStation.fromJson(model)));
+      List<Stop> nearbyStations = List<Stop>.from(stations.map((model) => Stop.fromJson(model)));
 
       return nearbyStations;
     }
@@ -83,7 +83,7 @@ class VbbApi {
 
       if (response.statusCode == 200) {
         Iterable stations = json.decode(response.body);
-        List<NearbyStation> nearbyStations = List<NearbyStation>.from(stations.map((model) => NearbyStation.fromJson(model)));
+        List<Stop> nearbyStations = List<Stop>.from(stations.map((model) => Stop.fromJson(model)));
 
         return nearbyStations;
       }
@@ -114,12 +114,10 @@ class VbbApi {
     );
 
     if (response.statusCode == 200) {
-      DeparturesAtStop departuresAtStop = DeparturesAtStop.fromJson(json.decode(response.body));
+      Map<String, dynamic> departuresAtStop = json.decode(response.body);
+      List<Departure> departures = List<Departure>.from(departuresAtStop['departures'].map((model) => Departure.fromJson(model)));
 
-      if (departuresAtStop.departures == null) {
-        throw Exception("Failed to get departures at stop.");
-      }
-      return departuresAtStop.departures!;
+      return departures;
     }
     else if (response.statusCode >= 500 && response.statusCode < 600) {
       if (!context.mounted) {
@@ -149,12 +147,10 @@ class VbbApi {
       );
 
       if (response.statusCode == 200) {
-        DeparturesAtStop departuresAtStop = DeparturesAtStop.fromJson(json.decode(response.body));
+        Map<String, dynamic> departuresAtStop = json.decode(response.body);
+        List<Departure> departures = List<Departure>.from(departuresAtStop['departures'].map((model) => Departure.fromJson(model)));
 
-        if (departuresAtStop.departures == null) {
-          throw Exception("Failed to get departures at stop.");
-        }
-        return departuresAtStop.departures!;
+        return departures;
       }
       else {
         throw Exception("Failed to get departures at stop.");
@@ -165,7 +161,7 @@ class VbbApi {
     }
   }
 
-  static Future<NearbyStation> getStationInfo(String stopId) async {
+  static Future<Stop> getStationInfo(String stopId) async {
     final String host = await _getMainApiHost();
 
     final response = await http.get(
@@ -179,12 +175,12 @@ class VbbApi {
       )
     );
 
-    NearbyStation info = NearbyStation.fromJson(json.decode(response.body));
+    Stop info = Stop.fromJson(json.decode(response.body));
 
     return info;
   }
 
-  static Future<List<NearbyStation>> getStations(String query, BuildContext context) async {
+  static Future<List<Stop>> getStations(String query, BuildContext context) async {
     final String host = await _getMainApiHost();
 
     final response = await http.get(
@@ -201,7 +197,7 @@ class VbbApi {
 
     if (response.statusCode == 200) {
       Map<String, dynamic> stationsRaw = json.decode(response.body);
-      List<NearbyStation> stations = [];
+      List<Stop> stations = [];
       List<String> usedIds = [];
 
       for (var v in stationsRaw.values) {
@@ -241,19 +237,19 @@ class VbbApi {
 
       if (response.statusCode == 200) {
         Map<String, dynamic> stationsRaw = json.decode(response.body);
-        List<NearbyStation> stations = [];
+        List<Stop> stationsOther = [];
         List<String> usedIds = [];
 
         for (var v in stationsRaw.values) {
           Station station = Station.fromJson(v);
           String id = station.id!.split(":")[2];
           if (!usedIds.contains(id)){
-            stations.add(await getStationInfo(id));
+            stationsOther.add(await getStationInfo(id));
             usedIds.add(id);
           }
         }
 
-        return stations;
+        return stationsOther;
       }
       else {
         throw Exception("Failed to get stops.");
